+++ mlflow/odahuflow/trainer/helpers/fs.py
@@ -1,0 +1,15 @@
+import os
+import shutil
+
+
+def copytree(src, dst):
+    """
+    Copy file tree from <src> location to <dst> location
+    """
+    for item in os.listdir(src):
+        s = os.path.join(src, item)
+        d = os.path.join(dst, item)
+        if os.path.isdir(s):
+            shutil.copytree(s, d)
+        else:
+            shutil.copy2(s, d)
