+++ mlflow/odahuflow/trainer/helpers/log.py
@@ -1,0 +1,12 @@
+import argparse
+import logging
+
+
+def setup_logging(args: argparse.Namespace) -> None:
+    """
+    Setup logging instance
+    """
+    log_level = logging.DEBUG if args.verbose else logging.INFO
+
+    logging.basicConfig(format='[odahuflow][%(levelname)5s] %(asctime)s - %(message)s', datefmt='%d-%b-%y %H:%M:%S',
+                        level=log_level)
