+++ mlflow/requirements.txt
@@ -1,4 +1,4 @@
 odahu-flow-sdk==1.3.0
 odahu-flow-cli==1.3.0
-mlflow==1.12.0
+mlflow>=1.13.0
 PyYAML>=3.1.2
