--- mlflow/tests/test_mlflow_runner.py
@@ -1,3 +1,0 @@
-
-def test_hello_world():
-    print('Ok!')
