+++ mlflow/version.py
@@ -1,0 +1,4 @@
+"""
+Version file
+"""
+__version__ = '1.3.0'
